move log and redis to commponents
