+++ components/redis/example_test.go
@@ -1,0 +1,2 @@
+package redis_test
+
